feat(knowledge-bases): add link-depth selector for URL sources (AYC-176)
