+++ ayunis-core-frontend/src/shared/locales/de/knowledge-bases.json
@@ -149,6 +149,11 @@
       "urlDialogTitle": "Website-URL hinzufügen",
       "urlDialogDescription": "Geben Sie die URL einer Website ein, um deren Inhalt in die Wissenssammlung zu importieren.",
       "urlLabel": "Website-URL",
+      "urlDepthLabel": "Linktiefe",
+      "urlDepthOption0": "Nur diese Seite",
+      "urlDepthOption1": "Diese Seite + verlinkte Seiten",
+      "urlDepthOption2": "Diese Seite + 2 Linkebenen",
+      "urlDepthHint": "Folgt Links auf derselben Website. Höhere Tiefen importieren mehr Seiten und dauern länger.",
       "addUrlSuccess": "URL erfolgreich hinzugefügt! Der Website-Inhalt wird verarbeitet.",
       "addUrlError": "URL konnte nicht hinzugefügt werden. Bitte überprüfen Sie die URL und versuchen Sie es erneut.",
       "addUrlRetrievalFailed": "Die Website konnte nicht erreicht werden. Bitte überprüfen Sie, ob die URL korrekt ist und die Website erreichbar ist.",
