+++ ayunis-core-frontend/src/shared/locales/en/knowledge-bases.json
@@ -149,6 +149,11 @@
       "urlDialogTitle": "Add Website URL",
       "urlDialogDescription": "Enter the URL of a website to import its content into the knowledge base.",
       "urlLabel": "Website URL",
+      "urlDepthLabel": "Link depth",
+      "urlDepthOption0": "Just this page",
+      "urlDepthOption1": "This page + linked pages",
+      "urlDepthOption2": "This page + 2 levels of links",
+      "urlDepthHint": "Follows links on the same website. Higher depths import more pages and take longer.",
       "addUrlSuccess": "URL added successfully! The website content is being processed.",
       "addUrlError": "Failed to add URL. Please check the URL and try again.",
       "addUrlRetrievalFailed": "The website could not be reached. Please check that the URL is correct and the website is accessible.",
